fix(republik): rename nameUtil to naming
