+++ packages/utils/index.js
@@ -2,5 +2,5 @@ module.exports = {
   applyPgInterval: require('./applyPgInterval'),
   ensureStringLength: require('./ensureStringLength'),
   paginate: require('./paginate'),
-  nameUtil: require('./nameUtil')
+  naming: require('./naming')
 }
